refac: move helper functions to utils
