+++ itconfig-macro/src/lib.rs
@@ -4,9 +4,11 @@
 mod ast;
 mod expand;
 mod parse;
+mod utils;
 
 extern crate proc_macro;
 extern crate proc_macro2;
+
 use self::proc_macro::TokenStream;
 use ast::RootNamespace;
 use quote::ToTokens;
